# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentOpinionOut": {
        "description": "Mirrors `shared.contracts.AgentOpinion` field-for-field.",
        "properties": {
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "concerns": {
            "items": {
              "type": "string"
            },
            "title": "Concerns",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "reasoning": {
            "title": "Reasoning",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/OpinionVerdict"
          }
        },
        "required": [
          "agent_name",
          "verdict",
          "reasoning",
          "concerns",
          "confidence"
        ],
        "title": "AgentOpinionOut",
        "type": "object"
      },
...
      "Direction": {
        "enum": [
          "INCREASE",
          "HOLD",
          "CLAWBACK"
        ],
        "title": "Direction",
        "type": "string"
      },
...
      "ErrorBody": {
        "description": "Every non-2xx response returns this shape as its JSON body.\n\n`code` is a short machine-readable slug (e.g. \"not_found\", \"forbidden\",\n\"validation_error\") for client branching; `message` is one sentence a human\nreads; `detail` carries whatever extra structure the specific failure needs\n(field errors, the resource id that was missing, etc.) and may be null.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpinionVerdict": {
        "description": "A single governance agent's stance on a recommendation, before opinions are\ncombined into one Recommendation.",
        "enum": [
          "CONCUR",
          "OBJECT",
          "ABSTAIN"
        ],
        "title": "OpinionVerdict",
        "type": "string"
      },
...
      "Page_RecommendationOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "1-indexed page number",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Items per page, as requested",
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matching records, independent of page size",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[RecommendationOut]",
        "type": "object"
      },
...
      "RecommendationDecision": {
        "description": "Request body for both `POST .../approve` and `POST .../reject`.\n\n`reason` is mandatory on both \u2014 approving *and* rejecting are mutations,\nand the cross-cutting rule doesn't carve out an exception for the\n\"yes\" path. Only ADMIN may call either endpoint (see `app/deps.py`,\n`Role.ADMIN` docstring: \"Reviewer explicitly cannot approve\").",
        "properties": {
          "reason": {
            "description": "Why this decision was made",
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RecommendationDecision",
        "type": "object"
      },
      "RecommendationOut": {
        "description": "Mirrors `shared.contracts.Recommendation` field-for-field.",
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "clamped": {
            "title": "Clamped",
            "type": "boolean"
          },
          "clamped_from": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clamped From"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "governance_mode": {
            "title": "Governance Mode",
            "type": "string"
          },
          "has_dissent": {
            "title": "Has Dissent",
            "type": "boolean"
          },
          "opinions": {
            "items": {
              "$ref": "#/components/schemas/AgentOpinionOut"
            },
            "title": "Opinions",
            "type": "array"
          },
          "proposed_limit": {
            "title": "Proposed Limit",
            "type": "integer"
          },
          "proposed_rung": {
            "title": "Proposed Rung",
            "type": "integer"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RecommendationStatus"
          },
          "trust_evaluation_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Evaluation Ref"
          }
        },
        "required": [
          "recommendation_id",
          "agent_id",
          "schema_version",
          "direction",
          "proposed_limit",
          "proposed_rung",
          "rationale",
          "opinions",
          "has_dissent",
          "confidence",
          "governance_mode",
          "status",
          "trust_evaluation_ref",
          "generated_at",
          "clamped",
          "clamped_from"
        ],
        "title": "RecommendationOut",
        "type": "object"
      },
      "RecommendationStatus": {
        "description": "Lifecycle of a governance recommendation, from proposed to resolved.",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "SUPERSEDED"
        ],
        "title": "RecommendationStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The integration point for all four lanes: the Policy Engine, persistence and audit, the HTTP contract, and the hard ceiling. See docs/SYSTEM-EXPLAINED.md for the full design.",
    "title": "Adaptive AI Governance Platform \u2014 Backend",
    "version": "1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/agents/{agent_id}/recommendations": {
      "post": {
        "description": "Generate a fresh governance recommendation for `agent_id`: recompute\nits `TrustEvaluation` from persisted decisions, run the governance panel\nover it (`GOVERNANCE_MODE`, default `stub`), clamp the panel's proposal to\nwhat the evidence actually supports, and persist trust evaluation,\nrecommendation, and audit entry \u2014 all in this one request's transaction\n(`app.deps.get_session`; see `app/services/governance.py`).",
        "operationId": "create_recommendation_api_v1_agents__agent_id__recommendations_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A downstream dependency (e.g. governance) could not serve this request."
          }
        },
        "summary": "Create Recommendation",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/v1/recommendations": {
      "get": {
        "description": "List recommendations, newest first.",
        "operationId": "list_recommendations_api_v1_recommendations_get",
        "parameters": [
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RecommendationOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/{rec_id}": {
      "get": {
        "description": "Fetch one recommendation by id, including its full opinion panel.",
        "operationId": "get_recommendation_api_v1_recommendations__rec_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rec_id",
            "required": true,
            "schema": {
              "title": "Rec Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recommendation",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/{rec_id}/approve": {
      "post": {
        "description": "Authorize a pending recommendation. ADMIN only.\n\nWrites an `approvals` row (`decided_by`, `verdict=APPROVED`, `reason`,\n`decided_at`), flips `Recommendation.status` to `APPROVED`, and \u2014 in the\nsame transaction \u2014 writes the new `policy_versions` row that actually\nchanges `agents.current_limit`/`current_rung`\n(docs/lanes/vp.md: \"Never update agents.current_limit without writing a\npolicy_versions row in the same transaction\").",
        "operationId": "approve_recommendation_api_v1_recommendations__rec_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "rec_id",
            "required": true,
            "schema": {
              "title": "Rec Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The current role may not perform this action."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The resource is not in a state this action applies to."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Recommendation",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/{rec_id}/reject": {
      "post": {
        "description": "Reject a pending recommendation. ADMIN only, same as approve.\n\nWrites an `approvals` row with `verdict=REJECTED` and flips\n`Recommendation.status` to `REJECTED`. No policy version is written \u2014\nthe agent's limit does not change.",
        "operationId": "reject_recommendation_api_v1_recommendations__rec_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "rec_id",
            "required": true,
            "schema": {
              "title": "Rec Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The current role may not perform this action."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The resource is not in a state this action applies to."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Recommendation",
        "tags": [
          "recommendations"
        ]
      }
    },
...
  }
}